+++ lib/sbmt/kafka_consumer/config/probes/liveness_probe.rb
@@ -7,6 +7,6 @@ class Sbmt::KafkaConsumer::Config::Probes::LivenessProbe < Dry::Struct
   attribute :path, Sbmt::KafkaConsumer::Types::Strict::String
     .optional
     .default("/liveness")
-  attribute :timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(10)
+  attribute :timeout, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(300)
   attribute :max_error_count, Sbmt::KafkaConsumer::Types::Coercible::Integer.optional.default(10)
 end